add OK button to LearningScreen and implement Hilt to project

+++ model/src/main/java/ir/hamid/model/W504.kt
@@ -13,5 +13,5 @@ data class W504(
     @ColumnInfo(name = "sample") val sample: String?,
     @ColumnInfo(name = "definition") val definition: String?,
     @ColumnInfo(name = "translate") val translate: String,
-    @ColumnInfo(name = "review") val review: Int?
+    @ColumnInfo(name = "review") val review: Long?
 )
